Refs #32036. Move seq fitting tests to relevant test files.
